+++ collections/test/dimensions.txt
@@ -1,0 +1,2 @@
+360
+617
